+++ src/modules/friend/friends.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
 import { FriendsRepository } from 'src/domain/interface/friend/friends.repository';
+import { FriendWriteService } from 'src/domain/services/friend/friend-write.service';
 import { FriendsService } from 'src/domain/services/friend/friends.service';
 import { FriendsPrismaRepository } from 'src/infrastructure/repositories/friend/friends.prisma.repository';
 import { GatheringParticipationModules } from 'src/modules/gathering/gathering-participation.module';
@@ -10,6 +11,7 @@ import { FriendsController } from 'src/presentation/controllers/friend/friends.c
   imports: [UsersModule, GatheringParticipationModules],
   controllers: [FriendsController],
   providers: [
+    FriendWriteService,
     FriendsService,
     { provide: FriendsRepository, useClass: FriendsPrismaRepository },
   ],
